refactor(Lotto): 패키지 경로 변경
- domain 패키지로 이동
